Initial WASM API to get performance clusters
This gets a basic performance tracking into the WASM API.

It's not great - I set a global default subscriber in initializer of
WasmPark which is iffy because now I can't have more than one
park per page (wasn't planning to, but it's smelly) and the code
for accessing clusters is ugly. But hey it's a good first stab.